pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/django_observability"]

[project]
name = "django-observability"
version = "0.1.0"
description = "Comprehensive OpenTelemetry observability for Django with traces, logs, metrics, and profiling"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
keywords = ["Django", "OpenTelemetry", "Observability", "Tracing", "Logging", "Metrics", "Celery"]
authors = [
    { name = "Adin Hodovic", email = "hodovicadin@gmail.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 5.2",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Logging",
]

dependencies = [
    "Django>=5.2",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-django>=0.41b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "structlog>=23.1.0",
    "django-structlog>=5.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
# Celery support
celery = [
    "celery>=5.0.0",
    "opentelemetry-instrumentation-celery>=0.41b0",
]

# Prometheus metrics (hybrid approach)
prometheus = [
    "django-prometheus>=2.3.0",
    "prometheus-client>=0.17.0",
]

# Profiling support (push mode - optional)
profiling = [
    "pyroscope-io>=0.8.0",  # Only needed for push mode; pull mode uses Pyroscope server scraping
]

# Database instrumentation
postgres = [
    "opentelemetry-instrumentation-psycopg2>=0.41b0",
]

# Cache instrumentation
redis = [
    "opentelemetry-instrumentation-redis>=0.41b0",
]

# HTTP client instrumentation
http = [
    "opentelemetry-instrumentation-requests>=0.41b0",
    "opentelemetry-instrumentation-urllib3>=0.41b0",
]

# Beautiful dev logs
dev-logging = [
    "rich>=13.0.0",
]

# All features
all = [
    "django-observability[celery,prometheus,profiling,postgres,redis,http,dev-logging]",
]

# Development dependencies
dev = [
    "ruff>=0.8.0",
    "tox>=4.25.0",
    "coverage>=7.8.0",
    "django-stubs>=5.1.3",
    "mypy>=1.15.0",
    "pylint>=3.3.6",
    "pylint-django>=2.6.1",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.0",
    "pytest-django>=4.11.0",
    "pytest-clarity>=1.0.1",
    "pytest-mock>=3.14.0",
    "celery>=5.0.0",
    "redis>=4.0.0",
    "opentelemetry-instrumentation-redis>=0.41b0",
    "opentelemetry-instrumentation-requests>=0.41b0",
    "opentelemetry-instrumentation-urllib3>=0.41b0",
    "requests>=2.0.0",
    "urllib3>=1.26.0",
]

[project.urls]
Homepage = "https://github.com/adinhodovic/django-observability"
Documentation = "https://github.com/adinhodovic/django-observability"
Repository = "https://github.com/adinhodovic/django-observability"
Issues = "https://github.com/adinhodovic/django-observability/issues"

[tool.pytest.ini_options]
addopts = "--ds=tests.settings --reuse-db"
python_files = "tests.py test_*.py"
norecursedirs = ".git */migrations/* */static/* venv .venv .virtualenv node_modules .tox"
markers = [
    "integration: marks tests as integration tests that require Docker stack (deselect with '-m \"not integration\"')",
]

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
    ".venv",
    "venv",
    "virtualenv",
    "migrations",
    "node_modules",
    ".tox",
    "examples",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["django_observability"]

[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main"]
exclude = [".venv", "migrations.*", "examples"]

[tool.django-stubs]
django_settings_module = "tests.settings"

[tool.pylint.'MASTER']
load-plugins = "pylint_django"
ignore = "manage.py"

[tool.pylint.'FORMAT']
max-line-length = 150

[tool.pylint.'MESSAGES CONTROL']
disable = "missing-docstring,invalid-name,logging-fstring-interpolation,fixme,duplicate-code"

[tool.pylint.'DESIGN']
max-parents = 13

[tool.coverage.run]
branch = true
source = ["django_observability"]
omit = [
    "./.venv/*",
    "*/migrations/*",
    "*/apps.py",
    "*/tests/*",
]

[tool.coverage.report]
fail_under = 70
